Added funcionality to create a pdf file
Implemented the functionality to create a .PDF file.
Implemented the funtionality to download the .pdf file created.

Possible improvments:
Remove some of the 255+ files added
(it's not important if it takes longer to work as long as it works,
right? We're a bit over schedule)
Also, the link may not work because there is no file to download.
One of us will have to figure how to link the method and the anchor.